+++ pex/compatibility.py
@@ -108,6 +108,7 @@ if PY3:
 
     from urllib.error import HTTPError as HTTPError
     from urllib.request import build_opener as build_opener
+    from urllib.request import FileHandler as FileHandler
     from urllib.request import HTTPSHandler as HTTPSHandler
     from urllib.request import ProxyHandler as ProxyHandler
     from urllib.request import Request as Request
@@ -115,6 +116,7 @@ else:
     import urlparse as urlparse
 
     from urllib2 import build_opener as build_opener
+    from urllib2 import FileHandler as FileHandler
     from urllib2 import HTTPError as HTTPError
     from urllib2 import HTTPSHandler as HTTPSHandler
     from urllib2 import ProxyHandler as ProxyHandler
